first version folders refactor
