+++ src/pages/landingPage/components/faqs/FAQsComponent.tsx
@@ -1,8 +1,8 @@
 import { useEffect } from "react";
 import { useSelector,useDispatch } from "react-redux";
-import { RootState } from "../../store/store";
-import FaqsModified from "./FaqsModified";
-import { apiCall,deleteApiFaqs } from "../../store/reducers/faqsCall/faqsReducer";
+import { RootState } from "../../../../store/store";
+import FaqsModified from "../../../backOffice/components/faqsAdminView/FaqsModified";
+import { apiCall,deleteApiFaqs } from "../../../../store/reducers/faqsCall/faqsReducer";
 
 
 const FAQs = () => {
@@ -14,7 +14,7 @@ const FAQs = () => {
   }
 
   interface createToken {
-    acces_token: String;  
+    acces_token: string;  
   }
 
   interface createFaqs {
@@ -38,7 +38,7 @@ const FAQs = () => {
 
   useEffect(()=>{
     apiCall(dispatch)
-  },[])
+  },[dispatch])
 
  
   return (
