+++ src/pages/landingPage/components/header/interfaces/interfaces.tsx
@@ -1,0 +1,14 @@
+export interface ChildComponentProps {
+    setIsDropdownEnterButton: React.Dispatch<React.SetStateAction<boolean>>;
+    setisDropdownCuenta: React.Dispatch<React.SetStateAction<boolean>>;
+  }
+
+ export interface RootState {
+    apiPostRegister: ApiPostRegisterState;   
+  }
+  
+  export interface ApiPostRegisterState {
+    messageError: string;
+    isLoadingMessageError: boolean;
+  }
+
